processRestRequest should always return an array
processRestRequest should always return an array or throw an exception. 
I had an issue where postmark returned a status code that was not in the list, this made processRestRequest return NULL. 
According to the documentation at the top of the function, this is not supposed to happen. 

And that is not something the DynamicResponseModel expects. 
The DynamicResponseModel __construct *requires* an array as parameter. 

To avoid these issues, the processRestRequest should throw an Exception. I've modified the code so it now includes a default: statement to catch all status codes we haven't defined. I've used the Exception handling of status code 422 to do this because this looks like a generic one. 

I've also removed the break; statements. There is no execution of code after a return or a throw statement.