Fixed Exception: incompatible JSON input.
The call `$client->sendEmailWithTemplate("sender@example.com","recipient@example.com",65464,[]);` with `$templateModel` as an empty array as  results in the exception below:

PostmarkException: Received valid, but incompatible JSON input.

The reason behind this is the api expects TemplateModel to be json object and not json array. PHP's json_encode converts empty array to empty json array.