Allow array syntax to `sendEmail()` or `sendEmailWithTemplate()` methods (#138)
Addresses issue raised in #137 

I also have a wiki update and will create an issue with a patch for this
since GitHub doesn't allow PRs on wikis :(

A couple notes:

- I wrote an extra Exception and validation check for this but removed
it since the library seems to leverage the Postmark API as the authority
for exceptions
- The `READ_SELENIUM_OPEN_TRACKING_TOKEN` was added to the example token
list due to an error I encountered when running PHPUnit tests locally

Thanks for the opportunity! This will be very useful to have in the
workflow on the library. It changes several API requests in my
application to one.

:1st_place_medal: